% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{run_simulation}
\alias{run_simulation}
\title{Run a simple dummy simulation for the specified duration.}
\usage{
run_simulation(duration = 100)
}
\arguments{
\item{duration}{Numeric. The length of time to run the simulation.}
}
\value{
list. Dummy simulation results.
}
\description{
Run a simple dummy simulation for the specified duration.
}
